move TaskStatistics.kt from entity package to utils package
