+++ app/src/main/java/com/amsterdam/cutetudee/domain/utils/TaskStatistics.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.cutetudee.domain.entity
+package com.amsterdam.cutetudee.domain.utils
 
 data class TaskStatistics(
     val totalTasks: Int,
